ui: bc: separate miner and peer-network code
